fix(provider/pihole): prevent redundant updates when multiple targets are available
When the source provides multiple targets to be created/updated in PiHole only the first is used. On a subsequent update that will cause the record to be removed and added back, even if it did not change and the same value is still present in the new target list. On a network with latency this creates interruptions.

Can be reverted when PiHole implements a way to create local DNS records with multiple IPs in their API.

Signed-off-by: Ion Marusic <ion.marusic@gmail.com>